enhanced logic for multithreading
